Moving upgrade logic to a handler
